Rename 'trees' directory to 'plants'
