+++ src/main/java/robosky/ether/world/feature/plants/CloudDaisyPatchFeature.java
@@ -1,0 +1,4 @@
+package robosky.ether.world.feature.plants;
+
+public class CloudDaisyPatchFeature {
+}
