+++ src/main/scala/robosky/ether/block/CloudDaisiesBlock.scala
@@ -8,7 +8,7 @@ import net.minecraft.tag.BlockTags
 import net.minecraft.util.math.BlockPos
 import net.minecraft.util.shape.VoxelShape
 import net.minecraft.world.BlockView
-import robosky.ether.world.feature.trees.EtherSaplingGenerator
+import robosky.ether.world.feature.plants.EtherSaplingGenerator
 
 object CloudDaisiesBlock {
     val SHAPE: VoxelShape =
