+++ src/main/scala/robosky/ether/world/feature/plants/EtherSaplingGenerator.scala
@@ -1,4 +1,4 @@
-package robosky.ether.world.feature.trees
+package robosky.ether.world.feature.plants
 
 import java.util.Random
